requirements.txt
aiogram==3.13.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pytz==2023.3
aiofiles==23.2.1

# База данных PostgreSQL
SQLAlchemy==2.0.41
asyncpg>=0.28,<1.0

# GigaChat и LangChain
langchain-gigachat==0.3.12
langchain-core==0.3.72
langchain-community==0.3.27
langchain-text-splitters==0.3.9

# Векторная база для RAG (опционально)
langchain-chroma==0.2.5
chromadb==1.0.15
# Зависимости для sentence-transformers (совместимые версии)
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
# torch и torchvision должны быть совместимыми версиями
# torch 2.2.1 совместим с torchvision 0.17.1
torch==2.2.1
torchvision==0.17.1
sentence-transformers>=2.2.0

# Дополнительные зависимости для ИИ
pymorphy3==2.0.4
pymorphy3-dicts-ru==2.4.417150.4580142
PyYAML==6.0.2
pypdf

# Планировщик задач
APScheduler==3.11.0

